unload minifig, rightclick tasks, required tools added
created unload minifig button and added functionality such that it stops
a raider's current task and drops whatever they are holding. made it so
that a raider that is selected can be given a task by rightclicking that
task, and raiders will not start a new task on their own when selected.
made it so that a raider must have the required tool to perform a task.