dedicatedResources drops when interrupting raider, clearTask sets position to nearest space, pathfinding returns all equal solutions
several bugs have been addressed and (hopefully) fixed in this commit.
When a raider that is bringing ore to a building site is stopped by the
player, the dedicatedResources number for that site is updated
appropriately, and the site is added back to tasksAvailable or
tasksUnavailable if necessary. raider now sets its space to the nearest
space in clearTask to fix some pathing inefficiency. calculatePath now
has the option to return all equal distance paths, and is now used in
conjunction with findClosestStartPath method so that a unit will choose
the shortest path that also has the closest start space to it. These
changes were implemented over a long period of time, and are in need of
more robust testing to ensure that they all work as intended without
causing any new bugs.