fixes for Raider stop task button, pathing to new task, selecting 'collect' tasks
Raider stop task button now resets 'grabPercent' when a Raider was
interrupted in the middle of picking up or dropping an object. Raider no
longer moves after choosing a new task if the Raider is already in range
of the task (also avoids situations where Raider updates its image angle
in an undesirable manner due to incredibly slight change in position).
Raiders can no longer share a 'collect' task. Assigning a Raider to
pickup a collectable that another Raider has already set out to pickup
will not do anything, and when clicking on overlapping collectables,
collectables that are already assigned to a Raider will be ignored, and
the remaining pool of collectables clicked on will be compared to find
the collectable with the lowest draw depth as usual.