fixed trying to kill building objects multiple times
buildings and buildingSites lists are no longer iterated over and killed
(as these are merely spaces, they are all killed when destroying
terrain, so trying to kill them again is a bad idea). This fixes bug
where buildingPlacer gets stuck in the top-left corner of the screen
after leaving a level and then loading another or the same one.