camera can no longer go out of bounds
gameLayer camera is now restricted to world bounds, meaning the camera
will stop if it tries to go past the first or last tile in either
dimension. gameLayer no longer draws a solid background color at the
start of each frame, as there is no way for it to be visible now.