breadth first task sweep added to checkChooseNewTask
when searching for a new task to perform automatically, if no high
priority tasks are found, the raider will call calculatePath with an
optional flag that searches for a path to a task that the raider can
complete, rather than a predetermined space.
checkChooseCloserEquivalentResource now only selects resources of the
same type.