Revert holding check when selecting task, methodize resource location selection
Reverted portions of canPerformTask and checkSetTask from previous
commit dealing with the case where the raider is holding an object.
Selecting a task to perform while holding an object is now contained in
the attemptSelectResourceLocation method, referred to both when
automatically selecting a new task while holding something, as well as
when picking up a collect task.