cleanup deprecated raider code; default to walk when unable to perform task
Cleaned up deprecated findClosest method in Raider class. Task
assignment now falls back to walk task for each selected raider that is
not carrying the necessary tool to perform the task (eg. right clicking
some rubble will cause any selected raider that is not carrying a shovel
to walk to that spot instead of sweeping it).