Fixed selecting non-primary building pieces; tweaked power path selection string
Selecting part of a building that is not the main Space for that
building now automatically changes the selection to the main Space.
Power paths, when selected, now use the selection string "power path"
regardless of whether they are a user-created power path, a building
power path, or a special building-specific power path.