valid spaces created during OL loading phase touched, selected hidden spaces hide level
When loading in objects from the OL map, any object that causes terrain
to change (currently, this only applies to building spaces) will check
for a touched neighbor; if one exists, touchAllAdjacentSpaces will be
called from that space (this can matter if that building would have
otherwise been inside a wall). Selected spaces that are hidden will now
longer reveal their current level if they are a building.