move holdingAngleDifference to Collectable class, fix holding list bugs
holdingAngleDifference is now stored in the Collectable class rather
than the raider class, in order to simplify holding multiple objects.
Fixed several instances of holding still being treated as an object
rather than a list (for now certain functionality will be hard-coded to
a list of length 1, and gradually reimplemented).